Prevent gap jumping while seeking.
Sometimes, when seeking backwards, the seek takes long enough for the playhead
to erroneously detect a gap and gap jump back to its previous position.
This only seems to happen when casting, presumably due to the additional
delay from message-passing.
This prevents gap jumps from happening while the video is seeking, and it should
solve the issue.

Closes #1061

Change-Id: I363c69428d84029a7808229c9c0e18eb691dd7f8