Fix URI extension parsing in HLS parser
Instead of directly splitting on dots, use the URI parser to deal with
URL parameters and hash parameters, then split only the path component
of the URI to extract an extension.

This required the update of all of our fake URLs used in testing, from
test://foo to test:/foo, so that the Uri parser would be able to see
"/foo" as a path, rather than "foo" as a domain.

Closes #1085

Change-Id: Ie7140b03ac689f721562b2537b83432f99667585