Fix URI map consistency in the HLS parser
When we update the manifest, we need to have consistent URIs in our
maps.  Before this change, some were the original URIs, while others
were the "response URI", the URI after redirection and scheme
resolution.

Because of the need for consistency across updates,
For now, for consistency across updates, we will use the original
URIs, without redirects.

Ultimately, we should update a playlist from the most recent response
URI, and segment URIs should always be relative to that.  TODOs have
been added.

Issue #740

Change-Id: If3c64a83e1649f845710c81ba0c4856bc8ead81f