Regression Test for Variant Recreation
Added tests for how we recreate variants from StreamDB instances.

This change does make createVariants public, but as it does not use any
internal state, there should be no reason to hide it especially as it allows
us to test it directly.

This change renames createVariants to recreateVariants to better represent
what it is doing.

Change-Id: I5a770d690e582d95b7f31a4b656f86a4bbfce7bb