Cancel text mediaState's update when not showing captions
When the user chooses not to show captions, we unload the text stream,
by canceling the update of text mediaState, and deleting the text
mediaState.

Since setupStreams_() is an async function, I'm adding a boolean
variable to mark the request of unloading text stream, so as to prevent
the race condition caused by quickly hiding and showing captions.

Closes #1058.

Change-Id: I554522b542bde5db037d74aa3a499d1ac0001151