Fix a null exception with CastReceiver.destroy
This causes the cast receiver to ignore further proxy events after
the receiver has been destroyed. This is to prevent errors after
destroy.
This is a nearly inconsequential bug; destroy is only called when
the cast receiver app is about to be unloaded, after all.
However, this change is necessary for integration tests to work.

Change-Id: I7bd9cb0f815e8f0989838c7f205e0475eb1b729b