Fix weird page-load race on IE 11
Sometimes, on IE 11, the demo app fails to load because "shaka" is
undefined when shakaDemo.init() is called.  This seems to have been
introduced in 51263249 when we started deferring script loading.
This bug has not appeared in a release version yet.

It seems that IE 11 sets the readyState on the document at the wrong
time when scripts are deferred.  Edge and Chrome both do the right
thing.

This came up during debugging while working on #1111, but this fix is
not part of the fix for #1111.

Change-Id: Ia42466acdced7ca67fed5a6e8270620b7d8dcd50