Only support unprefixed orientation API
Prefixed versions of the screen orientation API are not compatible
with the unprefixed version.  Until there is demand for prefixed
versions, only use the unprefixed version (Chrome & Firefox).
Prefixed versions should be supported through a polyfill, if at all.

This also fixes a bug in the demo app that caused the screen
orientation exception to be silently ignored.

Closes #1111

Change-Id: I9a6e963e6312983bc4bf629ecb2b8e1ad782c49c