Fix CastReceiver integration tests on Chromecast
Because we mocked out canDisplayType to return false, and because that
is actually used on Chromecast in the MediaSource polyfill, this
caused the receiver integration test to fail with
CONTENT_UNSUPPORTED_BY_BROWSER.  Because nothing was listening for
errors, this error was shown only as a timeout.

This fixes the mock and listens for errors.

Change-Id: I05a62c439666cfe1a282aa523a75e362e1674a4b