Adds support for mp4a.40.34 mp3 in HLS
When including mp3 files in an HLS stream, the officially supported
codec is mp4a.40.34, according to the HLS FAQ (https://goo.gl/11ZdLh).
This codec string, unfortunately, is not recognized by MSE
implementations on platforms other than Safari.
Therefore, this change omits that codec string.

This change also modifies the start time code in the parser to assume
that mp3 files have a start time of 0.

Closes #1210

Change-Id: I5ccad3f8713fdcafecceff923a4d16ce8fb3ace3