[ES6] Disallow uncompiled mode in non-ES6 browsers
As we convert Shaka Player to ES6, we will lose the ability to run the
demo app in uncompiled mode on older browsers, such as IE 11.

This detects ES6 syntax support and disables the link to uncompiled
mode in the demo app.  It also prevents the loader from choosing
uncompiled mode by default in these older browsers.

Issue #1157

Change-Id: I13e3a80669daaabd38a056f5242ee713f3d014cc