Adjust appendWindowStart to help rounding errors.
Since we set the appendWindowStart to exactly the start of the Period,
we can get rounding issues where the first frame of the next Period
is removed.  If we remove the first frame, then MSE will remove up to
the next keyframe, causing a large gap.  Avoid slight rounding errors
by adjusting appendWindowStart back a little.

Bug: 71808910
Change-Id: I933352441c34046db7bf777264eb1ee4225d55e1