Compensate for TS timestamp rollover at load
If we see that the timestamps of any HLS live stream are over the
rollover limit, we will offset other streams by that limit until they
and in the same range as the non-rollover streams.

For example, an MP4 stream has no rollover, but an associated
segmented VTT text stream will be mapped to MPEGTS timestamps which do
rollover.  So we offset the VTT timestamps until they are in the same
range as the rest of the content.

This does not handle the case where timestamps roll over during
playback.

Issue #1191

Change-Id: I53bd0c5925601f3a432d653ddfa8a2652db84d71