Upgrade and unfork jsdoc
This updates jsdoc to the latest version from npm and forks a copy of
the default jsdoc template.

In subsequent commits, we will reapply our customizations to the
template and to a jsdoc plugin.

This also fixes several bad annotations that the new jsdoc failed to
parse.

Backported to v2.3.x

Issue #1259

Change-Id: I00a47270ea4754e1c96c43ca900d5cf889ab72e6