Make net engine returns error when out of retries
When changing networking engine to use AbortableOperation, we had an
accidental regression. Now, when it runs out of retries, it rejects
with an error of "undefined" instead of with the actual error that was
causing it to retry.
This changes networking engine to save the last error and return with
that. It also modifies the unit tests to catch this problem.

Closes #1278

Change-Id: I8af9a81d7a8ad58d195a8e9f1af622ddbc3592f2