Restore addPartialStream method.
This method was removed during the Variants refactoring.  This method
uses jasmine's asymmetric matchers to only match some of the fields.
This is important for tests so they only test the relevant fields.

This also removes implicit reuse of stream IDs.  Stream IDs must be
unique and there is a new method for cases where the intent is to
use duplicate streams.

Change-Id: I185df5f0beb018169ff537abb952ee4ebbf49be6