Split Playhead responsibilities.
This splits some of the independent code from Playhead into several
new helper classes.  This allows Playhead to be simpler and easier to
understand.

This keeps the new behaviors and classes as private pieces of Playhead
to keep Playhead conceptually in-charge of handing the Playhead.  It
still has the same responsibilities, but the code is split into other
files.

Issue #1224

Backported to v2.3.x

Change-Id: Ia828f902ba9490d128f4ca9cb1e34119ec93f188