Clear old text stream when loading new text stream
In our previous fixes for issue #1058, we made it so that we
remove the text engine when not displaying captions.
However, we did not clear MediaSource's buffered text. Thus, when
loading a new text stream after re-enabling the captions, the old
buffered text would still be displayed. Text tracks, in VOD content,
can potentially be a single file for the entire video. Thus, when
changing text languages while closed captions were disabled, the
actual captions did not change.

This modifies streaming engine to clear buffered text, to fix that
problem. It also updates unit tests.

Closes #1293

Change-Id: Iddf2591101286474862b52579d4cf6b464e450fb