Fix track selection during "streaming" event
This turned out to be three issues layered on top of one another:

First, we would fire the "streaming" event too early, before our final
stage of filtering. This meant that for multicodec content, you might
be selecting a track which will be subsequently filtered out. If the
track was no longer available after filtering, we would effectively
ignore your selection. This bug would only be triggered by multicodec
content (VP9 & H.264, for example).

This change fixes the ordering so that filtering comes before the
"streaming" event is fired.  A unit regression test was added.

Second, there was a code path in which we could ignore your track
selection. This was very easy to trigger with the "streaming" event,
and in an uncompiled build, would result in a failed assertion.

This change corrects assumptions about when we can ignore deferred
switches in the Player.  We had not previously accounted for the case
where we were still setting up the first period, which is easy to
trigger with the "streaming" event.  An integration regression test
was added.

Third, once the second bug was fixed, we could fail some assertions in
StreamingEngine and still effectively ignore your track selection.
This is because we did not reset a flag in the Player that tracks the
StreamingEngine setup state and informs when we defer switches.

This change resets the flag when we tear down streaming.  An
integration regression test was added.

Closes #1119

Change-Id: I16be372cfaf4d592ca895e41ad04333bc0b2eeee