Sync offline code between v2.3.x and master
This syncs the backports of offline changes with the latest code on
master, both to fix some bad cherry-picks and to make future
cherry-picks easier.

This also introduces some ES6 updates, style changes, and karma test
changes as a side-effect of the sync.

This undoes offline changes related to AbortableOperation, which is
not in this branch.

Issue #1248