Loosen tolerance on CEA-708 captions test
On some platforms, the sample used in that test leads to a
buffer start time of 1.4; on other platforms, it leads
to a buffer start time of 1.4667333333333332.
Perhaps this is a result of the implementation-specific rounding
differences in different mediaSource implementations, or perhaps there
us a very small gap in the content that some platforms handle slightly
differently.

Either way, this change loosens the tolerance on the check. That buffer
start time check wasn't the important part of the test, anyway.

Fixes: 77471871
Change-Id: I866fafe671566cc7116e7c99222f75225cffd476