Moved Storage Out of Download Manager
Move the use of storage engine out of download manager and back into
storage. Storage creates and manages the life of storage engine, so
it is easier for it to handle all storage functions.

Now, the callback that is given to download manager is used to pass
the downloaded array buffer back to storage so that it can be saved
into storage engine.

This moves all uses of storage engine into storage which will make it
easier to replace with Storage Muxer.

Issue #1248

Change-Id: I555788478af09d1d79c3db7534431e1d9daf4e20