Update fake tracks in player unit tests
Fake tracks now have multiple audio channels and a more realistic
variant layout.  The tests that use these tracks have been updated to
be less brittle by not relying on specific array indices or ID values.

This is a prelude to adding new tests for the issues found in #1013.

Issue #1013

Change-Id: If091674ba7a2de29c77c81556415a855f85b18af