Create IndexedDB Storage Mechanism
Create the indexeddb storage mechanism which will allow us to
connect to indexeddb storage cells. As we don't have a V1 Cell
yet, this only connects V2 and V3 cells. However the code does
assume that V1 will be added and only omits connecting the cell
when it sees the V1 object stores indexeddb.

Issue #1248

Backported to v2.4.x

Change-Id: I7395b8a334cd68795aab02ff656a5b49d1f5742b