Check license server vs asset when parsing app URI
Before this change, asset=foo.mpd in the demo app URI would select an
asset from the list whose manifest URI was foo.mpd, without regard for
anything else in the demo app URI.  So a demo app URI like
asset=foo.mpd;license=bar would still cause the predefined asset with
foo.mpd to be selected, even if that asset was configured to use a
different license server.

This led to the demo's inability to construct a URI that showed off a
preconfigured demo asset with a different configuration.

Now, the license server (if present in the demo URI) will be compared
to the asset's configuration before selecting it.  Any change to the
license server will now be treated as preloading a "custom" asset.

Change-Id: Ife5a890fcb1eca3782570ce1f2066fd42a45fb97