Fix support for startTime of 0
During a recent refactor, startTime || null got introduced.
This would not only use null as a default, but would also convert 0
into null, which changes the meaning of 0 from "exactly timestamp 0"
to "default start position" (which is non-zero for live streams).

This fix adds a regression test that fails without the fix.

Change-Id: I01e1628df9d141697fdb7d9fb5e77a16ac3dab5b